+++ craft_application/services/project.py
@@ -16,14 +16,21 @@
 from __future__ import annotations
 
 import copy
+import datetime
 import os
 import pathlib
 from typing import TYPE_CHECKING, Any, Literal, cast, final
 
 import craft_parts
 import craft_platforms
+import distro_support
 import pydantic
 from craft_cli import emit
+from distro_support.errors import (
+    NoESMInfoError,
+    UnknownDistributionError,
+    UnknownVersionError,
+)
 
 from craft_application import errors, grammar, util
 from craft_application.models import Platform
@@ -579,3 +586,70 @@ class ProjectService(base.AppService):
                 reportable=False,
                 logpath_report=False,
             ) from None
+
+    @staticmethod
+    def _is_supported_or_esm_on(
+        *, base: craft_platforms.DistroBase, date: datetime.date
+    ) -> bool:
+        """Check if the given base is supported or in extended support on a date."""
+        support_range = distro_support.get_support_range(base.distribution, base.series)
+        if support_range.is_supported_on(date) or support_range.is_in_development_on(
+            date
+        ):
+            return True
+        try:
+            return support_range.is_esm_on(date)
+        except NoESMInfoError:
+            return False
+
+    def check_base_is_supported(self) -> None:
+        """Check that this project's base and build-base are supported.
+
+        This method assumes a single-base project. Applications that use multi-base
+        projects must override this in order to use it.
+
+        :raises: CraftValidationError if either is unsupported.
+        """
+        project = self.get()
+        if project.base is None:
+            raise RuntimeError("No base detected when getting support range.")
+        base = craft_platforms.DistroBase.from_str(project.base)
+        build_base: craft_platforms.DistroBase | None = None
+        if project.build_base:
+            build_base = craft_platforms.DistroBase.from_str(project.build_base)
+
+            if build_base.series == "devel":
+                build_base = None
+
+        today = datetime.date.today()
+
+        try:
+            base_is_supported = self._is_supported_or_esm_on(base=base, date=today)
+        except (UnknownDistributionError, UnknownVersionError) as error:
+            # If distro-support doesn't know about this base, assume it's supported.
+            emit.debug(str(error))
+            base_is_supported = True
+        if build_base is not None:
+            try:
+                build_base_is_supported = self._is_supported_or_esm_on(
+                    base=build_base, date=today
+                )
+            except (UnknownDistributionError, UnknownVersionError) as error:
+                # Likewise assume unknown build bases are supported.
+                emit.debug(str(error))
+                build_base_is_supported = True
+        else:
+            build_base_is_supported = base_is_supported
+
+        if base_is_supported and build_base_is_supported:
+            return
+        message = (
+            f"Base '{base}' has reached the end of its lifespan."
+            if not base_is_supported
+            else f"Build base '{build_base}' has reached the end of its lifespan."
+        )
+        raise errors.CraftValidationError(
+            message,
+            resolution="If you know the risks and want to continue, rerun with --old-bases.",
+            retcode=os.EX_DATAERR,
+        )