+++ docs/reference/changelog.rst
@@ -7,10 +7,19 @@ Changelog
 5.7.0 (2025-MM-DD)
 ------------------
 
+Application
+===========
+
+- Add a :py:attr:`~craft_application.application.AppMetadata.check_supported_base`
+  option to ``AppMetadata``, allowing the application to opt into checking that the
+  base is supported.
+
 Services
 ========
 
 - Add a new Proxy service that configures an instance to connect to a proxy.
+- The Lifecycle service now configures the overlay to use ``old-releases.ubuntu.com`` if
+  the release has been migrated to that domain.
 
 For a complete list of commits, check out the `5.7.0`_ release on GitHub.